+++ test.js
@@ -1,9 +1,9 @@
-import * as convert from 'xml-js'
-import * as path from 'path'
-import * as fs from 'fs'
-import * as wmts from './'
-import { Format, BBox } from './'
+const convert = require('xml-js')
+const path = require('path')
+const fs = require('fs')
+const wmts = require('.')
 
+// Variables
 const title = 'Tile Service'
 const identifier = 'tile-service-123'
 const abstract = '© OSM data'
@@ -11,8 +11,8 @@ const minzoom = 10
 const maxzoom = 18
 const url = 'http://localhost:80/WMTS'
 const keywords = ['world', 'imagery', 'wmts']
-const format: Format = 'jpg'
-const bbox: BBox = [-180, -85, 180, 85]
+const format = 'jpg'
+const bbox = [-180, -85, 180, 85]
 const spaces = 2
 const options = {
   title,
@@ -24,7 +24,7 @@ const options = {
   bbox,
   url,
   keywords,
-  format,
+  format
 }
 
 /**
@@ -33,10 +33,10 @@ const options = {
  * @param {ElementCompact} json
  * @param {string} fixture
  */
-function compare(data: any, fixture: string) {
+function compare (data, fixture) {
   fixture = path.join('fixtures', fixture)
   let xml = data
-  if (typeof(data) !== 'string') { xml = convert.js2xml(data, {compact: true, spaces}) }
+  if (typeof (data) !== 'string') { xml = convert.js2xml(data, {compact: true, spaces}) }
   if (process.env.REGEN) { fs.writeFileSync(fixture, xml, 'utf-8') }
   expect(xml).toBe(fs.readFileSync(fixture, 'utf-8'))
 }
@@ -49,4 +49,14 @@ describe('Capabilities', () => {
   test('OperationsMetadata', () => compare(wmts.OperationsMetadata(url), 'OperationsMetadata.xml'))
   test('Layer', () => compare(wmts.Layer(options), 'Layer.xml'))
   test('getCapabilities', () => compare(wmts.getCapabilities(options), 'WMTSCapabilities.xml'))
+  test('throw error', () => expect(() => wmts.getCapabilities('invalid')).toThrow())
+})
+
+describe('utils', () => {
+  test('range(3)', () => expect(wmts.range(3)).toEqual([0, 1, 2]))
+  test('range(0, 3)', () => expect(wmts.range(0, 3)).toEqual([0, 1, 2]))
+  test('range(3, 0, -1)', () => expect(wmts.range(3, 0, -1)).toEqual([3, 2, 1]))
+  test('clean({foo: 10, bar: undefined})', () => expect(wmts.clean({foo: 10, bar: undefined})).toEqual({foo: 10}))
+  test('clean({foo: undefined, bar: undefined})', () => expect(wmts.clean({foo: undefined, bar: undefined})).toEqual({}))
+  test('clean({foo: 0})', () => expect(wmts.clean({foo: 0})).toEqual({foo: 0}))
 })